added return code for pending

+++ kernel/h/core/tick.h
@@ -18,7 +18,7 @@ typedef struct wdog {
 
 extern void wdog_init   (wdog_t * wd);
 extern void wdog_destroy(wdog_t * wd);
-extern void wdog_start  (wdog_t * wd, usize ticks, void * proc,
+extern void wdog_start  (wdog_t * wd, int ticks, void * proc,
                          void * a1, void * a2, void * a3, void * a4);
 extern void wdog_cancel (wdog_t * wd);
